Add ClientCert and ClientKey for newly required grpc auth (#16)
This diff adds 3 new options to be able to use dcrctl against a wallet that 
has authtype=clientcert

WalletGRPC should be set to true in this case and then it will pull the 
required cert and key fromthe default location (~/.dcrctl/client.pem and 
~/.dcrctl/client-key.pem) or use whatever path the usersets for ClientCert
and ClientKey. The client.pem needs to also be appended to the dcrwallet
clients.pem file.