Cleanup authtype option.
This modifies the config parsing for the new authtype option to enforce
only the two supported types and sets the default to the basic type.

It also enforces that the new authtype of clientcert requires TLS
regardless of the wallet flag being set because it is perfectly
acceptable to manually specify a wallet as a target without the wallet
flag that tells it to use the default ports.

Finally, it modifies it so that the client cert and key are loaded when
the authorization type is clientcert regardless of the wallet flag for
the same reason as above.