docs: Update README.md for Go 1.19 and rel 1.7.4.
This updates the README.md to require Go 1.18 or Go 1.19 and also
reworks the building section to be more accurate and explicitly specify
the release version because it's not necessarily the same as the latest
module version.

It also updates the building and isntalling section to be more
consistent with other repos.